Do not try to do content-based text replacement of __UND__ on directories. It seems to work, although we need a new directory added in a template to fully test it.
git-svn-id: https://gtsvn.uit.no/langtech/trunk/gtcore@82313 c7155fb1-f0a7-4240-a2fc-2600b6f42f90

+++ scripts/replace-dummy-langcode.sh
@@ -45,8 +45,8 @@ file=$3
 
 # Replace placeholder language code with real language code
 # In either a single file:
-if ! test "x$file" == "x"; then
-	if test -e $file ; then
+if test "x$file" != "x"; then
+	if test -e $file -a ! -d $file ; then
 	    cp $file $file~
 	    $SED -e "s/__UND__/$2/g" \
 	    	-e "s/__UND2C__/$iso2code/g" \
